third_party/WebKit/LayoutTests/inspector/modules-load-network.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script>
async function test()
{
    var initialModules = InspectorTest.loadedModules();
    await UI.inspectorView.panel("network");
    InspectorTest.dumpLoadedModules(initialModules);
    InspectorTest.completeTest();
}
</script>
</head>

<body onload="runTest()">
<p>This test validates set of loaded modules for Network panel.</p>
</body>
</html>
