third_party/WebKit/LayoutTests/inspector/modules-load-network-expected.txt

This test validates set of loaded modules for Network panel.

Loaded modules:
    cm
    cookie_table
    data_grid
    diff
    formatter
    network
    network_priorities
    object_ui
    perf_ui
    source_frame
    text_editor
    workspace_diff
